rewrite Gather to optimize use of strings and containers
Declaring Unbounded_String in a Visitor routine is not the best idea,
because it is a controlled type, i.e. is allocated on heap and needs to be
both initialized and finalized.

Also, Contains followed by Insert on a map requires an extra search, which
can be simply avoided by Insert variant that sets an "Inserted" status.