Add statement handling exception fallback
If an exception occured during exception handling we'd
halt compilation. With this commit, what we do instead is
report the error and continue on.

The result of the compilation will be ultimately not very useful,
but at least with this we can find out all the places where something
is going wrong rather easily