Rewrite do_array_assignment
Before the RHS was translated as an expression, creating temporary array for
every concatenation pair. Now it is as described in the comment preceding the
function body.

Add type of irep array

The RHS of array assignment may contain one or more concatenated arrays. Instead
of enforcing that a single irep is returned we want to return as many ireps as
there are operands of the concatenation, thus an array.