Fix array struct initialisation
Symex is not field-sensitive yet.

Create a temporary struct and assign it in one swoop to Array_Temp - so that
Symex does not see the struct as having been changed after its creation and can
therefore see it as constant - which means that the struct member that refers to
"allocated size" remains visible/accessible.