Add support for modular types with power-of-2 boundaries
Adds initial support for modular types. If the boundary is a power of 2 we can
represent the mod type as an unsignedbv without any special handling, but the
boundary value wouldn't fit into that unsignedbv, hence why we're handling this
case specially. The "not power of 2" case requires special arithmetic operators
and explicit tracking of the boundary value, which is why it's going to be
handled seperately in the future.