Handle private_type_declarations
The Full View of the private_type_declaration is obtained
from the tree on encountering the declaration, processed and
and added to the global symbol table.  When the corresponding
full_type_declaration is found a check is made if it has
a Partial or Private View and, if so, the declaration has already
been processed and no more has to be done.