Handle deferred_constant_declarations
The completion of a deferred_constant_declaraton is processed on
encountering the deferred_constant_declaration unless the declaration
has the pragma Import applied in which case it has no completion.
If the completion is processed the constant is registered in the
global symbol table so that it is not reprocessed when the completion,
the full_constant_delaration is encountered.