Introduce the support for symbol bounds
There are three bounds type now: Nat, Real, Symbol. The bounded types contain
two numbers: index of the lower bound and index of the upper bound. These
indices are unique and shared among all three bound types. Therefore we need to
also remember which index represents a bound of which type. Hence the
all_bounds_table is a vector of bound types (and the index of the bound is the
index inside this vector). Then each bound type has a map: index ->
actual_bound that stores the value of the bound.