Process a subprogram's parameters before processing its body
PR 220 introduced a regression.  It entered subprogram parameters into
the symbol table before processing their declaration causing the
processing of a subprogram specification to fail.
This could happen if a subprogram body was declared without a
separate declaration (perfectly legal to do this in Ada).

This change removes the code from PR 220 and ensures a
subprogram specification is processed before its body.