Generalise handling n_in
in Do_Expression. Before we simply gave up if the rhs node did not contain
low/high bound. Now we unroll the type until defining-identifier and extract the
scalar-range. Seems to be more generic.

also

Generalise handling enum parameters so that we only replace the constants (those
behind a symbol expression and the symbol table containing the value). Enabling
the test.