Make New_Irep private
In the future, new code should always use the Make_* family of
functions.

This will make very generic Irep code *slightly* harder to write, but
outside of the Ireps module itself it shouldn't really ever be
necessary, and generally speaking code that relies less on stateful
setters and instead constructs objects in a valid state to begin with
will be easier to follow.