Make Source_Location Irep again
Previously we just mapped Source_Location to GNATs Source_Ptr type. This works
well enough if we just want to present a literal source location, however CBMC
source locations can carry extra informations (for instance a comment which is
used for display of assertion messages) which can't be represent by Source_Ptr.
Now that Source_Locations are regular Ireps again we can add this information in
the places that can make use of this.