Add 'pretty' assertion condition to assertions
This currently reads the actual file and does some miniscule amount of parsing
to extract the actual source code. I had considered trying a pretty printer, but
the one in gnat right now seems to be quite limited and hard to use, and getting
the output exactly the way we want would be a bit hard using that, and writing
another pretty printer from scratch would be a pretty monumental task. This
approach looks a bit hacky, but should work in all cases thanks to Ada's syntax
being very regular.

Right now there's a bit of a performance issue in that we actually have to scan
through the file several times every time we see a '-' character; This can be
avoided by either using a different File API or by restructuring the code if
needed (I'd consider this version more of a PoC, although in practice I expect
the performance impact will be tolerable even on fairly large projects).