Handled GNAT BUG DETECTED reports that get truncated by gnat2goto crashes
There seems to be a couple of cases where gnat2goto fails with an
assertion failure in the middle of outputing the GNAT BUG DETECTED
report - meaning it never gets to print the ending deliminator. The new
script handles this better so it actually spots these cases and reports
them.