Some extra ignores
With the following justification: there are 4 classes of features we don't
support.

1. No support at all.
2. Is not relevant to the kind of analysis we are doing or the phase of
compilation and building that we are working on.
3. Support is not necessary to capture the executable semantics of the program.
Is actually supported in the translation.

We also have a number of options of what to do:

A. Give an unsupported feature and die.
B. Give an unsupported feature and continue.
C. Give some other kind of warning.
D. Silently ignore.
E. Actually implement it.

We propose the following mapping:

1 -> B if possible, A. if not.
2 -> D.
3 -> ??? Not clear, C or D.
4 -> E.

Breakdown of specific features touch in this PR.

These we will need to support but we need to do that in linking not in the
direct translation:

- Suppress initialization : an unsupported feature but not a fatal error -> B
- Initializes : an unsupported feature but not a fatal error -> B
- Linker options : an unsupported feature but not a fatal error -> B

The following are 3. because abstract state is not part of the compiled program.
Maybe, at some point in the future, we might want to improve the tooling to use
these but it is not incorrect for us to just ignore them:

- Abstract state : ignore for now
- N_Abstract_Subprogram_Declaration : Abstract subprogram declaration
- N_Private_Type_Declaration : Abstract type declaration unsupported

These are 2.'s so D:

- Ada 05 : ignore
- Ada 2012 : Ignore
- No strict aliasing : ignore
- Unreferenced : ignore
- Elaborate body : ignore
- N_Freeze_Generic_Entity
- N_Freeze_Entity

This is kind of a pain because I don't think we should need to support this but
... we may for complicated reasons, so B:

- Atomic : an unsupported feature but not a fatal error