Updated test for checking the withing of a subprogram body more than once.
The same subprogram body is withed by two different compilation units.
It is important that the subprogram spec is only processed once and that
its body is processed, but only once.
The test.py is modified to specify the main program so that the
other subprogram bodies do not have a a goto start function added.