Update test script to handle the generic body files and extra spec files.
A generic body does not generate a json_symtab file, so need to
check for their existence.
A library level generic instantiation is a declaration - a spec file
but in AST it is replaced by its instance.  Gnat2goto needs to process
such spec files.  They can be added using the list parameter specs in
the call to prove in test.py.