Update golden results.
Main differences are due to:
Now reporting an unsupported node for multi-language imports/exports
rather than a warning message.

Pragma/aspect Global now acceptrd as a no-op for ASVAT analysis but
meaningful for ASVAT modelling.

In StratoX a failed precondition in arrays.ads seems to be replaced by
an assertion failure in atree.adb.  The precondition in arrays.ads is
not changed by this mod, so the changes made to the code must now avoid
this particular precondition failure but the code then fails later
in atree.adb.