Fixed problems arising from applying new array model to latest master.
Most problems due missing range checks.

Also a misplaced statement in Handle_Parameter.

A strange problem to do with the use of Internal_Source_Location.
The variable containing the Irep for the Internal_Source_Location is
somehow being corrupted.  At present have not found where or how this
corruption is occuring.  Fixed, for the moment by regenerating the
Internal_Source_Location Irep each time the function is called.

Updated the regression test results.