Fix remaining problems with Assertion at line 479 range_check.adb.
Removed existing assertion because there can be a diiference in the
types of the lower and upper bounds, especially if one of the bounds
is a literal value (one bound may be a bounded bv type whereas the
other may be an unbounded bv type).

Moved all the necessary conversions and checks to Make_Range_Expression
and introduced some new utility routines in goto_utils.