Update golden results.
Extra reports of unknown size from ASVAT size model - probably
due to look up of components which have a null node as they are
associated with an unsupported declaration.

One worrying extra failure in libkeccak of a stack overflow.
I think it stems from a unit test which was not being analysed before.