This change addresses issue 314.
It removes the bottleneck of creating a large string containing
the text version of the JSON Symbol before outputting the string
to the json_symtab file.

A JASON_Value Write procedure has been created in a child package
of GNATCOLL.JASON (GNATCOLL.JSON.File) which outputs the text version
of a JSON_Value directly to a given text file.