The size of subprogram parameters was not being registered.
The size of subprogram parameters of a definite subtype were not
being registered in the ASVAT.Size_Model.
Parameters of an indefinite type are still not registered because
the size of an indefinite type is undefined.
Gnat2goto issues an unsupported message if 'Size is applied to an
parameter of an indefinite type and returns the value 0.